+++ express-mission-prisma/src/middlewares/RoleDescription
@@ -5,3 +5,7 @@
 들어가는 파일:
 authMiddleware.js: 사용자가 로그인했는지, 권한이 있는지 확인.
 errorHandler.js: 요청 처리 중 발생한 에러를 처리.
+
+
+
+말 그대로 미들웨어. 음...일종의 컴포넌트?를 정의하는 곳
