+++ express-mission-prisma/src/repositories/RoleDescription
@@ -6,3 +6,6 @@
 들어가는 파일:
 userRepository.js: 사용자 정보를 데이터베이스에서 가져오거나 저장.
 productRepository.js: 상품 정보를 데이터베이스에서 관리.
+
+
+프리즈마. 즉, 데이터베이스의 데이터를 추가, 수정 등 동작을하는 함수를 정의한는 곳
