+++ express-mission-prisma/src/services/RoleDescription
@@ -5,3 +5,7 @@
 들어가는 파일:
 userService.js: 사용자 관련 작업 처리 (로그인, 회원가입 등).
 productService.js: 상품 관련 작업 처리 (상품 등록, 수정 등).
+
+
+
+내부기능. 즉, 미들웨어나 라우트에서 사용할 함수를 정의하는 곳
